requirements.txt
flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
mlxtend>=0.22.0
tensorflow>=2.13.0
requests>=2.31.0
python-dateutil>=2.8.0
geopy>=2.3.0
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
python-dotenv>=1.0.0
Pillow>=9.5.0
opencv-python-headless>=4.7.0
google-generativeai>=0.3.0
Werkzeug>=2.3.0
gunicorn>=20.1.0 